pyproject.toml
[project]
name = "hsnu-copilt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord>=2.3",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.11.0",
    "google-generativeai>=0.8.5",
    "google-search-results>=2.4.2",
    "googlemaps>=4.10.0",
    "sentence-transformers>=4.1.0",
]
